requirements.txt
flask==3.0.2
flask-wtf==1.2.1
wtforms==3.1.2
email-validator==2.1.0.post1
werkzeug==3.0.1
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
requests==2.31.0
python-dotenv==1.0.0
mailersend==0.1.0 